Move JSFingerprintExtractor code to FingerprintExtractor
